Initial Capsicum work.
This is the initial work to apply FreeBSD's Capsicum to Tor. Note that
there's still a whole lot of cleanup, error checking/handling, etc. that
needs to happen. I haven't checked to ensure compiling works on Linux,
so things are probably broken there.

This work is far from over. Lots of improvements need to be made before
it's even sent upstream. This is not production-quality code.

*IMPORTANT*
Since libcrypto's X509 verification code eventually calls into
gmtime(3) to perform notBefore and notAfter validation, enabling Sandbox
mode on a Capsicum-enabled system could potentially enable an attacker
to present an expired or not-yet-valid certificate, which Tor might
accept. gmtime(3) would be unable to properly convert between local
system time and GMT, thereby opening a window in which certificates with
an improper notBefore or notAfter could be accepted. This issue is being
researched and hopefully a solution developed.